refactor(chat): live item -> live directory
